requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
dask[complete]>=2023.5.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0

# AWS SDK
boto3>=1.28.0
botocore>=1.31.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Dashboard
streamlit>=1.25.0
plotly>=5.15.0
folium>=0.14.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
holidays>=0.30

# Logging & Monitoring
python-json-logger>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Data Validation
great-expectations>=0.17.0

# Development
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
